Changes to preferred name behaviour:
- Rename username to preferred name
- Remove initial "set username" screen
- Some other copy changes
