+++ app/controllers/application_controller.rb
@@ -10,7 +10,7 @@ class ApplicationController < ActionController::Base
   end
 
   def after_sign_in_path_for(user)
-    stored_location_for(user) || login_dashboard_path(user)
+    stored_location_for(user) || dashboard_path
   end
 
   def after_sign_out_path_for(_user)
@@ -18,14 +18,6 @@ class ApplicationController < ActionController::Base
   end
 
 protected
-
-  def login_dashboard_path(user)
-    if user.account_created
-      dashboard_path
-    else
-      new_username_path
-    end
-  end
 
   def release_version
     ENV["RELEASE_VERSION"] || "-"
