+++ app/views/preferred_name/edit.html.erb
@@ -1,0 +1,18 @@
+<% content_for :before_content, govuk_back_link(
+  text: "Back",
+  href: dashboard_path)
+%>
+
+<div class="govuk-grid-row">
+  <div class="govuk-grid-column-two-thirds">
+    <h1 class="govuk-heading-xl">Edit your name</h1>
+
+    <%= form_with model: current_user, url: preferred_name_path do |f| %>
+      <%= f.govuk_text_field :preferred_name,
+        label: { text: "What is your preferred name?" },
+        hint: { text: "For example, Jess rather than Jessica." },
+        width: "two-thirds" %>
+      <%= f.govuk_submit "Save" %>
+    <% end %>
+  </div>
+</div>
