--- app/views/username/new.html.erb
@@ -1,15 +1,0 @@
-<% content_for :before_content, govuk_back_link(
-  text: "Back",
-  href: dashboard_path)
-%>
-
-<div class="govuk-grid-row">
-  <div class="govuk-grid-column-two-thirds">
-    <h1 class="govuk-heading-xl">Create a username</h1>
-
-    <%= form_with model: current_user, method: :post, url: username_path do |f| %>
-      <%= f.govuk_text_field :username, label: { text: "Username" }, width: "two-thirds" %>
-      <%= f.govuk_submit "Continue" %>
-    <% end %>
-  </div>
-</div>
