+++ db/migrate/20210504100552_rename_username.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+class RenameUsername < ActiveRecord::Migration[6.1]
+  def change
+    rename_column :users, :username, :preferred_name
+  end
+end
