+++ db/seeds/cip_seed_dump.rb
@@ -704,3 +704,6 @@ MentorMaterial.import([:id, :title, :content, :core_induction_programme_id, :cou
   ["2af78f3c-1d68-4772-8c73-63fc7d457a9d", "Challenge", "**Mentor handbook | B9 | Challenge<br />**\n\n**Study<br />**\n\nGet yourself into a strong position to mentor your teacher by working through the following:<br />\n\n|                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                          |\n| -------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- |\n| **Key takeaways for this module<br />**Your teacher can support pupils to be more open to challenge if they understand:<br />- Pupil motivation is driven by intrinsic and extrinsic factors, prior experiences and perceptions of success.<br /> - Teachers who give pupils experiences of success build not only motivation but also resilience and belief in their ability to succeed.<br /> - Teachers with knowledge of how pupils learn can better balance challenge, support and promote pupil success, which makes them open to challenge.<br /> |\n\nCLICK TO WATCH THE VIDEO B9<br />\n\nREAD THE EVIDENCE SUMMARY<br />\n\n# **Teaching challenge:<br />**\n\nWhile pupils in Ms Sterling’s class are generally keen to give tasks a go, they often take the safe option and seek support from their teacher or peers whenever possible. How can Ms Sterling help pupils to adopt behaviours that make them more open to challenge?**Key idea:**\n\nPupils who experience success are more likely to be motivated, resilient and open to challenge.<br />\n\n**The role of success<br />**\n\nPart of Ms Sterling’s job is to maximise pupil learning – this means providing the right level of challenge. However, as well as offering challenge, it is also important that tasks\\_ \\_enable pupils to experience a high success rate. This balance is a tricky one to strike.<br />\n\nEvidence suggests that if pupils struggle but are ultimately successful with a task, it is more likely they will remember the material (EEF, 2017). Where pupils have experienced success, they are likely to put in more effort, be more motivated and show more confidence in the future (Coe et al, 2014). This is because where pupils believe in their abilities to complete a specific task, they are more persistent at that task. Their investment is driven by their perceptions of success and failure, particularly if they have limited experience of meaningful success in the past, for example if they are low attaining pupils (Gutman & Schoon, 2013).<br />\n\n**Establishing high expectations of success<br />**\n\nResearch suggests that teachers should aim for pupils to be successful around 80% of the time (Rosenshine, 2012). Ms Sterling can build pupil expectations that they will succeed in a task by:<br />\n\n- **Offering rewards and praise:**Providing extrinsic motivation when pupils attempt challenging work. Using positive reinforcement more than negative works best (IES, 2008).<br />\n- **Attribution:**Linking effort and success for pupils when introducing or framing tasks (Coe et al., 2014).<br />\n- **Avoiding lavish praise:**If used\\*\\*\\*\\*without merit, praise can lower pupil confidence in their own ability (Coe et al., 2014).<br />\n- **Championing challenge** Sharing _how_ task effort will lead to future learning success. For example, directing pupils’ attention to others who have succeeded due to their effort, or saying: “this work is designed to be tough, don’t worry if you struggle, this will help you learn.”<br />\n\nThese strategies rely on a combination of intrinsic and extrinsic motivation. It is worth noting that while teachers can harness \\_extrinsic \\_motivation to help get pupils started, \\_intrinsic \\_motivation is likely to get pupils to stick at tasks, particularly when things get tricky (Lazowski & Hulleman, 2016).<br />\n\n**Effective teaching raises success rates<br />**\n\nMs Sterling can increase the chance pupils succeed at challenging tasks by using her emerging expertise in:<br />\n\n- How pupils learn (Deans for Impact, 2015), for example taking care not to overload their working memories.<br />\n- Her phase or subject specialism (Rosenshine, 2012; Coe et al., 2014), for example ensuring pupils have had enough input before they attempt challenging tasks, especially with specific barriers they might experience if the topic is particularly tricky, or they have special educational needs.<br />\n\nHer classroom climate is also crucial: when Ms Sterling sets challenging work, there will be times when pupils fail. Building a classroom where pupils trust that failure is okay is therefore important to help pupils deal with failure as a natural part of learning.<br />\n\n**Nuances and caveats** <br />\n\nThe relationship between teacher expectations and pupil outcomes is indirect. Teachers can best convey high expectations by getting the balance of challenge and support right. This will ensure pupils experience success, which should increase their motivation and sense of self-worth, also supporting their resilience (Coe et al., 2014).<br />\n\nWhen engineering a high success rate for pupils, Ms Sterling must be careful not to remove challenge altogether. Setting pupils up for success in unchallenging tasks does not build motivation and can embed low expectations if pupils interpret this as low teacher expectations (Coe et al., 2014). Instead, Ms Sterling should ensure she provides enough scaffolding for pupils to be successful and withdraw the scaffolding as pupils get better at a task (Rosenshine, 2012). She can also explain why she is withdrawing scaffolding.\\*\\*\\*\\*\n\nPupil success at a task is an indication\\_ \\_that they have successfully learnt lesson content, but not a sure sign – ‘learning’ and ‘performance’ are different.<br />\n\nSELECT A DEVELOPMENT AREA\\*\\*\\*\\*\n\n**Consider the development areas for this module (below). Then make a note of the area you plan to zoom in on and when you plan to visit so you can observe your teacher in this area. Familiarise yourself with the focused development areas. You will select one later when you observe your teacher.<br />**\n\n|                             |                                                                                                                                                                                                                                                                                                                                                                                                                                                                           |\n| --------------------------- | ------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- |\n| Development areas<br />     | Focused development areas<br />                                                                                                                                                                                                                                                                                                                                                                                                                                           |\n| Building complexity<br />   | - Teacher sets tasks that build in complexity and supports pupils to see this challenge as positive and surmountable.<br /> - Teacher asks questions that challenge pupils when their basic knowledge is secure.<br />                                                                                                                                                                                                                                                    |\n| Championing challenge<br /> | - Teacher models the behaviours and mindsets involved in tackling challenging tasks and how to embrace and overcome errors.<br /> - **Teacher promotes challenge using consistent and intentional language, so pupils see it as desirable and understand how it benefits their learning.<br />** - Teacher praises behaviours pupils use to tackle challenging tasks.<br /> - Teacher helps build a sense of momentum and pace in the class by fostering challenge.<br /> |\n| Removing scaffolding<br />  | - Teacher explains the importance of removing scaffolds and models how to gradually do so when knowledge is secure, and pupils are becoming more successful.<br /> - Teacher praises pupils who, when they have secure knowledge, challenge themselves by removing scaffolds.<br /> - Teacher gives pupils concrete and specific ways they can monitor their progress so they know when they should gradually remove scaffolds themselves.<br />                          |\n\nEXAMPLES OF PRECISE TARGETS<br />\n\n|                                             |                                                                                                                                                                                                                                                                                                                             |\n| ------------------------------------------- | --------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- |\n| If your teacher is..<br />                  | Then your precise target might be…<br />                                                                                                                                                                                                                                                                                    |\n| Not doing it at all…<br />                  | Use consistent language to describe the benefits of challenging tasks and link this to the benefits to pupils' learning.<br />                                                                                                                                                                                              |\n| Doing it but needs some improvement…<br />  | When pupils are working on tasks they find challenging, remind pupils of the progress they have made through persevering with challenging tasks, e.g. \"Last term you were doing ...; you have pushed yourselves, and now you are able to...\"<br />                                                                          |\n| Doing it well and needs some stretch…<br /> | During or following a task that pupils found challenging, plan in reflection time that causes pupils to think about the progress they have made through tackling the challenging task. Encourage pupils to use the same positive, intentional language you have used with them to describe the benefits of challenge.<br /> |\n\n**Observe<br />**\n\nConsider the following questions based on a short (approximately 15 minute) observation of your teacher.<br />\n\n**What was your teacher’s previous target? Are they meeting it? How do you know?<br />**\n\n**Thinking about the development area you have selected for this module, what is your teacher already doing well in this area? Which focused development area best aligns with what your teacher needs to get better at? What one precise target (bite-sized action) might you work with them on during your mentor meeting?**\n\nREMINDER: You can choose to stick with this previous target if they have not made enough progress. When moving on to a new precise target, you can select one from the table above or, if this doesn’t fit your teacher’s needs, you can write your own.<br />\n\n**How will you model the target to your teacher to show them what good looks like? What questions will you ask to check your teacher understands the model? For example, ‘How it is different from your current practice?’ and ‘What impact might it have on your practice and pupils?’<br />**\n\n**Reminder:** Your model should help your teacher develop their ability in some of the following:<br />\n\n- Set tasks that stretch pupils, but which are achievable, within a challenging curriculum.<br />\n- Remove scaffolding only when pupils are achieving a high degree of success in applying previously taught material.<br />\n- Plan activities around what you want pupils to think hard about.<br />\n- Support pupils to master challenging content which builds towards long-term goals.<br />\n- Help pupils to journey from needing extrinsic motivation to being motivated to work intrinsically.<br />\n\nNext, meet with your teacher to work through the ‘feedback’ stage of instructional coaching. <br />\n\n**References<br />**\n\nCoe, R., Aloisi, C., Higgins, S., & Major, L. E. (2014). _What makes great teaching. Review of the underpinning research_. Durham University. bit.ly/ecf-coe.<br />\n\n<br />\n\nDeans for Impact (2015). *The Science of Learning.* [bit.ly/ecf-dea](https://deansforimpact.org/resources/the-science-of-learning/).  <br />\n\n<br />\n\nEducation Endowment Foundation (2017). *Metacognition and Self-regulated learning Guidance Report.* bit.ly/ecf-eef   <br />\n\n<br />\n\nGutman, L. & Schoon, L. (2013). *The impact of non-cognitive skills on the outcomes of young people.*  <br />\n\n<br />\n\nInstitute of Education Sciences (2008). *Reducing Behavior Problems in the Elementary School Classroom.* bit.ly/ecf-ies.  <br />\n\n<br />\n\nLazowski, R. A., & Hulleman, C. S. (2016). Motivation Interventions in Education: A Meta-Analytic Review. *Review of Educational Research*, *86*(2), 602–640. <br />\n\n<br />\n\nRosenshine, B. (2012) Principles of Instruction: Research-based strategies that all teachers should know. *American Educator*, 12–20. bit.ly/ecf-ros.<br />\n", "483d68d3-aae6-416c-b9ec-6103fe2cd067", "a4dc302c-ab71-4d7b-a10a-3116a778e8d5", "ac3dc47f-e7fd-41b5-8c37-f096e4be451e", "cfd61715-2975-4853-b3e8-fd7a2f40a0b7"],
   ["c75a4d94-eb4e-43d4-98ab-00986e980f17", "Independent practice", "**Mentor handbook | B10 | Independent practice<br />**\n\n**Study<br />**\n\nGet yourself into a strong position to mentor your teacher by working through the following:<br />\n\n|                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                               |\n| ------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- |\n| **Key takeaways for this module<br />**Your teacher can promote behaviours that support independent practice if they understand that:<br />- Pupils need to understand the benefits of practice to their long-term learning, even if it feels hard.<br /> - To practise independently, pupils need enough support and clear behavioural expectations. Teachers also need to check pupil understanding of support and expectations.<br /> - Pupils need to be held to account to practise independently.<br /> |\n\nCLICK TO WATCH THE VIDEO B10<br />\n\nREAD THE EVIDENCE SUMMARY<br />\n\n# **Teaching challenge:<br />**\n\nMs Silva feels that her pupils are now more open to challenge and will sometimes try to work without scaffolding. However, they still prefer to work with teacher or peer support and may be reluctant to work on their own. Ms Silva finds pupils engage well in helping her solve a tricky problem on the board, but when she says, “your turn”, there are some blank faces. How can she get pupils to adopt behaviours that support independent practice?<br />\n\n# **Key idea:<br />**\n\nRegular, purposeful practice is vital for pupil learning, so teachers need to develop routines and behaviours that support independent practice.<br />\n\n**The benefits and challenges of getting pupils to practise**\n\nIndependent practice is vital to pupil learning and success. To learn, pupils need to think hard about the content they have been taught (Coe, 2013). Effective teachers give plenty of class time for independent practice (Rosenshine, 2012). It has clear benefits for pupils learning in terms of:<br />\n\n- **Developing pupil fluency:** When pupils practise, their understanding becomes more fluent and automatic, making it easier for pupils to apply their knowledge and learn new material (Rosenshine, 2012).<br />\n- **Helping pupils remember:**For example, retrieval practice (getting pupils to recall what they have learned) is one of the best ways to ensure pupils remember learning at a later date (Pashler et al., 2008).\\*\\*\\*\\*\n\nSometimes pupils avoid thinking hard – we all do. They also form unrealistic views of how much they know. Therefore, when given a choice, they often don’t choose effective study approaches (Pashler et al., 2008). For example, re-reading their notes may feel easy. Trying to recall what they have learned without support from peers or scaffolding feels harder, but is far more effective (Dunlosky et al., 2013). As pupils are also easily distracted, Ms Silva can best ensure they think hard by insisting that pupils practise independently.<br />\n\n**Getting pupils ready for independent practice<br />**\n\nMs Silva should ask herself whether pupils are ready to practise independently. Do they realise why the effort of independent practice is important for their learning? Explaining the benefits of independent practice will help. For instance, she could explain that:<br />\n\n- We learn what we think hard about.<br />\n- Less support leads to better learning once pupils are ready to practise.<br />\n- Effort makes success more likely (Coe et al., 2014).<br />\n\nShe also needs ensure pupils practise successfully, as failure can damage pupil motivation and sense of self-worth (Coe et al., 2014). Independent practice is best done alone, so Ms Silva needs to provide enough support to ensure success. For example:<br />\n\n- Introducing a manageable amount of new material.<br />\n- Leading teacher-guided practice on the same material pupils will practise independently.<br />\n- Providing scaffolding (Rosenshine, 2012).<br />\n\n**Clear expectations and routines enhance independent practice<br />**\n\nMs Silva can set up independent practice consistently, in a way that develops routines over time. This is likely to contribute to pupil success, helping pupils to value practice. So, Ms Silva should consider how she will consistently:<br />\n\n- **Set clear behavioural and task expectations:** (Coe et al., 2014) This means outlining the behaviours she expects to see during independent practice: ”I should see everyone focusing on their own work silently”, and the task and support she expects pupils to use: ”I want you to complete this exercise on the worksheet, without looking at the work we did last week”.<br />\n- **Check for understanding:** When introducing the independent practice tasks, teachers should ask specific, task-focused questions to get a clear sense of whether pupils have understood instructions (Rosenshine, 2012). Ms Silva should avoid questions like: “do we all understand this?”, where pupils’ default answer is ”yes”, even if they may not understand, or social pressure prevents them admitting to gaps in understanding (Rosenshine, 2012).<br />\n- **Circulate:** Checking pupils are following instructions and holding them to account is distinct from supporting with work (Lemov, 2015). Research suggests that as teachers circulate, they should check in with individual pupils for no more than around 30 seconds (Rosenshine, 2012). Longer contacts could disrupt pupil independence by suggesting that teacher support is available.<br />\n\nIf Ms Silva is finding many questions arise during independent practice, it might be that pupils are not ready or that they do not understand expectations. She might consider stopping practice, checking that enough support is in place and that pupils have understood her expectations.<br />\n\n**Nuances and caveats** <br />\n\nTeachers should not set independent tasks when pupils have very little knowledge of a topic (Coe et al., 2014). Pupils will need to be built up and supported through teacher input first. Once this has happened, relevant homework can also be good independent practice of what has been learned, particularly for older pupils. For young pupils, playful practice can be led by pupil interest and teachers should provide just enough support for pupils to be successful (Deans for Impact, 2019).<br />\n\nEffective monitoring during independent practice is often non-verbal – for example, the teacher standing and visibly scanning the classroom. While the importance of reinforcing behaviours with public praise is well known (IES, 2008), during independent practice there is a risk of this distracting pupil attention.<br />\n\nThere are also benefits of collaborative learning (Kirschner et al., 2018; Rosenshine, 2012), however Ms Silva might prioritise getting independent practice right first to make it more likely collaborative practice succeeds.<br />\n\nSELECT A DEVELOPMENT AREA<br />\n\n**Consider the development areas for this module (below). Then make a note of the area you plan to zoom in on and when you plan to visit so you can observe your teacher in this area. Familiarise yourself with the focused development areas. You will select one later when you observe your teacher.<br />**\n\n|                                                                |                                                                                                                                                                                                                                                                                                                                                                                                                                                                               |\n| -------------------------------------------------------------- | ----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- |\n| Development areas<br />                                        | Focused development areas<br />                                                                                                                                                                                                                                                                                                                                                                                                                                               |\n| Establishing clear expectations for independent practice<br /> | - Teacher explains to pupils the importance of independent practice and their behavioural expectations for independent practice.<br /> - Teacher ensures pupils know what to do during independent practice as well as the conditions to work in.<br /> - Teacher supports pupils to turn the expectations for independent practice into a routine by keeping certain expectations the same and having pupils practise them.<br />                                            |\n| Ensuring pupils understand expectations<br />                  | - **Teacher checks that pupils understand the instructions and behavioural expectations before they begin independent practice.<br />** - Teacher models to pupils the behaviours and steps to take during independent practice when they think pupils may struggle.<br />                                                                                                                                                                                                    |\n| Circulate to check<br />                                       | - Teacher is clear on the behaviours they need to check for when circulating the classroom and phrases these as visible and positive behaviours.<br /> - Teacher pre-emptively circulates to check pupils are meeting their expectations and acknowledges where pupils are doing so.<br /> - Teacher supports pupils to meet behaviour expectations, e.g. using non-verbal reminders for a pupil or a whole class reset if there are lots of pupils who need reminding.<br /> |\n\nEXAMPLES OF PRECISE TARGETS<br />\n\n|                                             |                                                                                                                                                                        |\n| ------------------------------------------- | ---------------------------------------------------------------------------------------------------------------------------------------------------------------------- |\n| If your teacher is..<br />                  | Then your precise target might be…<br />                                                                                                                               |\n| Not doing it at all…<br />                  | Ask pupils to repeat back the instructions for the independent task to check they have understood what to do.<br />                                                    |\n| Doing it but needs some improvement…<br />  | Ask pupils to repeat back the instructions for the independent task including the behaviours that are expected of them and what to do if they are stuck.<br />         |\n| Doing it well and needs some stretch…<br /> | If a pupil asked cannot tell you the instructions for the task or behaviour expectations, have a peer state them and return to the initial pupil to repeat them.<br /> |\n\n**Observe<br />**\n\nConsider the following questions based on a short (approximately 15 minute) observation of your teacher.<br />\n\n**What was your teacher’s previous target? Are they meeting it? How do you know?<br />**\n\n- **Thinking about the development area you have selected for this module, what is your teacher already doing well in this area? Which focused development area best aligns with what your teacher needs to get better at? What one precise target (bite-sized action) might you work with them on during your mentor meeting?<br />**\n- REMINDER: You can choose to stick with this previous target if they have not made enough progress. When moving on to a new precise target, you can select one from the table above or, if this doesn’t fit your teacher’s needs, you can write your own.<br />\n\n**How will you model the target to your teacher to show them what good looks like? What questions will you ask to check your teacher understands the model? For example, ‘How it is different from your current practice?’ and ‘What impact might it have on your practice and pupils?’<br />**\n\n**Reminder:** Your model should help your teacher develop their ability in some of the following:<br />\n\n- Teach and rigorously maintain clear behavioural expectations for independent work.<br />\n- Monitor pupil work during lessons.<br />\n- Give manageable, specific and sequential instructions.<br />\n- Check pupils’ understanding of instructions before a task begins.<br />\n- Create and explicitly teach routines in line with the school ethos that maximise time for learning.<br />\n- Reinforce routines.<br />\n\nNext, meet with your teacher to work through the ‘Feedback’ stage of instructional coaching. <br />\n\n**References<br />**\n\nCoe, R. (2013). *Improving Education: A triumph of hope over experience*. Centre for Evaluation and Monitoring.  [bit.ly/ecf-coe](https://www.suttontrust.com/wp-content/uploads/2014/10/What-Makes-Great-Teaching-REPORT.pdf)2.<br />\n\n<br />\n\nCoe, R., Aloisi, C., Higgins, S., & Major, L. E. (2014). _What makes great teaching. Review of the underpinning research_. Durham University. [bit.ly/ecf-coe](https://www.suttontrust.com/wp-content/uploads/2014/10/What-Makes-Great-Teaching-REPORT.pdf).<br />\n\n<br />\n\nDunlosky, J., Rawson, K. A., Marsh, E. J., Nathan, M. J., & Willingham, D. T. (2013) _Improving students’ learning with effective learning techniques: Promising directions from cognitive and educational psychology._ Psychological Science in the Public Interest, Supplement, 14(1), 4–58.<br />\n\nInstitute of Education Sciences (2008). *Reducing Behavior Problems in the Elementary School Classroom.* bit.ly/ecf-ies.  <br />\n\n<br />\n\nKirschner, P., Sweller, J., Kirschner, F. & Zambrano, J. (2018). From cognitive load theory to collaborative cognitive load theory. In *International Journal of Computer-Supported Collaborative Learning*, 13(2), 213-233.  <br />\n\n<br />\n\nLemov, D. (2015). *Teach Like a Champion 2.0*.  Jossey-Bass. 2nd ed. Edition. <br />\n\n<br />\n\nPashler, H., McDaniel, M., Rohrer, D., & Bjork, R. (2008). Learning Styles: Concepts and Evidence. *Psychological Science in the Public Interest*, *9 *(3).  bit.ly/ecf-pas.<br />\n\n<br />\n\nRosenshine, B. (2012) Principles of Instruction: Research-based strategies that all teachers should know. *American Educator*, 12–20. bit.ly/ecf-ros<br />\n", "483d68d3-aae6-416c-b9ec-6103fe2cd067", "a4dc302c-ab71-4d7b-a10a-3116a778e8d5", "ac3dc47f-e7fd-41b5-8c37-f096e4be451e", "55494017-528d-44fc-9189-1b182d477e6b"]
 ])
+CoreInductionProgramme.import([:id, :name, :course_year_one_id, :course_year_two_id], [
+  ["57482e25-2654-4c54-9219-8583d3336d1e", "Test Core induction programme", nil, nil]
+])
