+++ spec/cypress/support/step_definitions/common-navigation.js
@@ -6,8 +6,7 @@ const pagePaths = {
   cookie: "/cookies",
   start: "/",
   dashboard: "/dashboard",
-  "create username": "/username/new",
-  "edit username": "/username/edit",
+  "edit preferred name": "/preferred-name/edit",
   "sign in": "/users/sign_in",
   "core induction programme index": "/core-induction-programmes",
   "core induction programme year": "/core-induction-programmes/:id",
